requirements.txt
jsonclasses>=3.1.3,<4.0.0
jsonclasses-pymongo>=3.1.3,<4.0.0
jsonclasses-server>=3.1.3,<4.0.0
inflection>=0.5.1,<0.6.0
flask[async]>=2.0.1,<2.1.0
python-dotenv>=0.19.1,<0.20.0
bcrypt>=3.2.0,<4.0.0
pyjwt>=2.1.0,<3.0.0
